+++ mvn/utils/misc.py
@@ -114,7 +114,7 @@ def is_master():
 
 
 def drop_na(x):
-    return np.float32(list(filter(
+    return np.float64(list(filter(
         lambda x: np.isfinite(x), x
     )))
