+++ post/utils.py
@@ -9,7 +9,7 @@ def parse_epochs(epochs):
     losses = {}
 
     for key in loss_keys:
-        losses[key] = np.float32([
+        losses[key] = np.float64([
             np.mean(epoch[key])
             for epoch in epochs
         ])
